feat: implement the Redis adapter
We had to make a few changes to the redis module (in the vendor/
directory), in order to implement the feature. We will remove it once
the changes are merged upstream.

See also: https://github.com/denodrivers/redis/issues/335

This implementation is compatible with the Node.js one (same format for
the Redis channels and the packets).

See also: https://github.com/socketio/socket.io-redis-adapter