fix(engine): properly pause the polling transport during upgrade
Before that change, the client was stuck waiting for the ongoing HTTP
long-polling request to be cleanly closed by the server, and the
upgrade process would fail after the 10s delay.

Note: the Node.js client implementation does not resume polling
("unpause") in that case, which looks like a bug.

Related:

- https://github.com/socketio/socket.io-deno/issues/4
- https://github.com/socketio/engine.io/commit/00f973829214b45cbdcde735a547887c1b1bcf21
- https://github.com/socketio/engine.io/commit/1c96ca4fe8bc2345f81cb4caca92bb77cc331791
- https://github.com/socketio/engine.io-client/blob/dfee8ded722a8c4f9f773505d0c77b4561569863/lib/transports/polling.ts#L84-L116